Professional/Insurance_Data_Pipeline/Layer_Silver/pyproject.toml
[project]
name = "GenerateData"
version = "1.0.0"
description = "A project for generating synthetic healthcare data"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.5.3",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pyyaml>=6.0.2",
]
